requirements.txt
Django~=4.2.0
djangorestframework~=3.15.1